+++ src/Application/Query/Badge/UserBadgeByEventView.php
@@ -57,6 +57,12 @@ class UserBadgeByEventView
     /** @var bool */
     public $isRightImageFullHeight;
 
+    /** @var string|null */
+    public $headerLeftColor;
+
+    /** @var string|null */
+    public $headerRightColor;
+
     public function __construct(
         ?string $sheetTitle,
         ?string $firstName,
@@ -72,7 +78,9 @@ class UserBadgeByEventView
         bool $isMirrored,
         ?string $leftImage,
         ?string $rightImage,
-        bool $isRightImageFullHeight
+        bool $isRightImageFullHeight,
+        ?string $headerLeftColor,
+        ?string $headerRightColor
     ) {
         $this->sheetTitle = $sheetTitle;
         $this->firstName = $firstName;
@@ -89,5 +97,7 @@ class UserBadgeByEventView
         $this->leftImage = $leftImage;
         $this->rightImage = $rightImage;
         $this->isRightImageFullHeight = $isRightImageFullHeight;
+        $this->headerLeftColor = $headerLeftColor;
+        $this->headerRightColor = $headerRightColor;
     }
 }